Move snapshot session modules into session boundary
Keep snapshot policy and persistence beside the existing session ownership helpers while leaving the large session manager move for a separate PR.

Constraint: structural cleanup must stay reviewable and avoid repository-wide line-ending churn

Rejected: moving session-manager in the same pass | its broad import surface made the diff too noisy for a safe cleanup PR

Confidence: high

Scope-risk: narrow

Directive: move session-manager only after deciding whether to preserve a compatibility shim or normalize affected CRLF files deliberately

Tested: npm run build; npm run lint; npm run lint:tier; git diff --check; npm test -- --runTestsByPath tests/session-state-persistence.test.ts tests/tools/session-snapshot-policy.test.ts tests/src/self-healing-wiring.test.ts tests/integration/idle-adaptive-monitoring.test.ts --runInBand --silent

Not-tested: full Jest suite
