+++ src/self-healing.ts
@@ -5,8 +5,8 @@
 export { ChromeProcessWatchdog } from './chrome/process-watchdog';
 export type { ProcessWatchdogOptions, ProcessWatchdogEvents } from './chrome/process-watchdog';
 
-export { SessionStatePersistence } from './session-state-persistence';
-export type { PersistedSessionState, PersistedSession, PersistedWorker, PersistedTarget } from './session-state-persistence';
+export { SessionStatePersistence } from './session/state-persistence';
+export type { PersistedSessionState, PersistedSession, PersistedWorker, PersistedTarget } from './session/state-persistence';
 
 export { TabHealthMonitor } from './cdp/tab-health-monitor';
 export type { TabHealthMonitorOptions, TabHealthInfo, TabHealthStatus } from './cdp/tab-health-monitor';
@@ -23,7 +23,7 @@ export {
   normalizeAutoSnapshotPolicy,
   redactSnapshotText,
   shouldTakeAutoSnapshot,
-} from './session-snapshot-policy';
+} from './session/snapshot-policy';
 export type {
   AutoSnapshotPolicyConfig,
   NormalizedAutoSnapshotPolicy,
@@ -32,4 +32,4 @@ export type {
   SnapshotMemoInput,
   SnapshotPolicyState,
   SnapshotTrigger,
-} from './session-snapshot-policy';
+} from './session/snapshot-policy';
